mongo_setup_google_emb/requirements.txt
streamlit
pandas
pymongo
rouge-score
bert-score
groq
python-dotenv
numpy
torch
transformers